Use builder images in dockerfiles
The final tippecanoe image doesn't need the compiler in it. By
using a builder image, we can install the compiler and build
tippecanoe, then copy the results into a smaller final image.

In Travis, run a build that stops on the builder image stage
so that make test can be run in the script stage. In normal
usage for a user that's not necessary. Update the travis image
to bionic, as trusty is now in extended maint - and the only
thing these jobs are doing is running docker.

While we're in there, update centos to centos8 and ubuntu to 20.04,
but add build args that allow overriding that if desired.

Finally, ubuntu doesn't need build-essential, which pulls in all
of the tools needed to build debian packages. Like centos it just
needs gcc, g++ and make.